feat: improve the View Model for authentication
Added a `UserState` class that defines a user state: either `Loading`, `Success` or `Error`.

Added an observable user state `userState` to the class.
Improved `signUpWithEmail` function to take a context as argument (used to save the access token in Android's shared preferences).
Added `onSuccess` and `onFailure` in the call to the Model's `login` method.