test: removed when condition in NoAlertDialog
The test coverage was under 80% because of a when condition in the
NoAlertDialog composable. This condition was in charge of modifying the
text displayed depending on which in which tab there were no alerts.

For the moment, the UI has only placeholder values and is only there to
mock how it will look once the ViewModel and Model for the alert list are finished.
Without them, it is not possible to test all the situations (when they
are no alerts, when there are alerts). So I decided to remove this functionality until the
Model and ViewModel are implemented.