fix: avoid submitting an alert in Alert screen if invalid
Checked for a location before submitting the alert.

Checked for product and urgency level before submitting: I used states (`productIsSelected` and `urgencyIsSelected`) and their setting function (`setProductIsSelected` and `setUrgencyIsSelected`).

Clicking on the submission button with an invalid alert now shows a `Toast` with corresponding error message and does not navigate anywhere.