refactor: cleanup the code in `AlertList.kt`
Extracted relevant values to constant variables.
Added a private `enum` class `AlertListsTab` to represent the two tabs `MY_ALERTS` and `PALS_ALERTS`. Now, instead of having a `selectedTabIndex` with an unnamed, difficult to understand value, we can just pass the `AlertListsTab` around.
Changed variables names for better maintainability and modularity, and so that the code is self-documenting.
Added documentation for functions and TODOs where needed.