refactor: clean up code in `AuthenticationComponents.kt`
Modified signatures to have optional instead of mandatory parameters for the `GradedBackground`, and `AuthenticationWelcomeText` components. The parameters passed being the same nearly every time, it's cleaner like this.
Created new object `AuthenticationScreens` in `C.Tag` that now contains `SignInScreen` and `SignUpScreen`, as well as some shared authentication components. This is so that I could get rid of the `testTag` parameter in some of the components.
Re-moved `AuthenticationGoogleButton` to `SignIn.kt` as other authentication screens don't use it.