feat: update `CreateProfile` screen to integrate user view model
Updated the `onClick` of the "Save" button to call the view model and attempt to save the user to the database. If not successful, shows toast.

Extracted the `onClick` actions to an auxiliary function `attemptSaveUserData`.

Updated `MainActivity` to initialise the user view model with a Supabase user model, and pass it as parameter to the `CreateProfile` screen.