test: fix and add tests for the location dropdown in Create Alert screen
Added a second testTag on the dropdown item in the form of the content description to be able to test them.
Added tests to check that the location dropdown works as expected:
 - when empty location suggestions list (no suggestions or locationVM failure), no dropdown item is shown
 - when location suggestions list has < 4 items, items are shown
 - when location suggestions list has >= 4 items, only three are shown
Moved `when`s and `setContent` from the `setUp` to every test to be able to choose the `when` return value per-test.